+++ src/__tests__/schema.ts
@@ -14,15 +14,18 @@ import {
   String,
   ManyToOne,
   Mixin,
-  Pk,
   OneToOne,
   Map,
   Key,
   Float,
   Raw,
+  Fields,
+  References,
+  OnUpdate,
+  OnDelete,
 } from '../';
 
-// from: https://www.prisma.io/docs/concepts/components/prisma-schema#example
+// roughly from: https://www.prisma.io/docs/concepts/components/prisma-schema#example
 
 const Role = Enum(
   'Role',
@@ -33,6 +36,7 @@ const Role = Enum(
 
 const Post = Model('Post');
 const User = Model('User');
+const Star = Model('Star');
 
 const Timestamps = Mixin()
   .Field('createdAt', DateTime(Default('now()')))
@@ -40,24 +44,56 @@ const Timestamps = Mixin()
 
 // prettier-ignore
 User
-  .Field("id",          Int(Id, Default("autoincrement()"), Map("_id"), Raw("@db.Value('foo')")))
-  .Field("email",       String(Unique))
-  .Field("name",        String(Nullable))
-  .Field("height",      Float(Default(1.80)))
-  .Field("role",        Role("USER", Nullable))
-  .Relation("posts",    OneToMany(Post))
-  .Field("bestPostId",  Int())
-  .Relation("bestPost", OneToOne(Post, Pk("id").Fk("bestPostId")))
+  .Field('id',          Int(Id, Default('autoincrement()'), Map('_id'), Raw('@db.Value(\'foo\')')))
+  .Field('email',       String(Unique))
+  .Field('name',        String(Nullable))
+  .Field('height',      Float(Default(1.80)))
+  .Field('role',        Role('USER', Nullable))
+  .Relation('posts',    OneToMany(Post, "WrittenPosts"))
+  .Relation('pinned',   OneToOne(Post, "PinnedPost", Nullable))
   .Mixin(Timestamps);
 
 // prettier-ignore
 Post
-  .Field("id",          Int(Id, Default("autoincrement()")))
-  .Field("published",   Boolean(Default(false)))
-  .Field("title",       String(Limit(255)))
-  .Field("authorId",    Int(Nullable))
-  .Relation("author",   ManyToOne(User, Pk("id").Fk("authorId"), Nullable))
+  .Field('id',          Int(Id, Default('autoincrement()')))
+  .Field('published',   Boolean(Default(false)))
+  .Field('title',       String(Limit(255)))
+  .Field('authorId',    Int(Nullable))
+  .Relation('author',   ManyToOne(User, "WrittenPosts", Fields('authorId'), References('id'), OnUpdate("Restrict"), OnDelete("SetNull"), Nullable))
+  .Field('pinnedById',  Int(Nullable))
+  .Relation('pinnedBy', OneToOne(User, "PinnedPost", Fields('pinnedById'), References('id'), Nullable))
+  .Relation('stars',    OneToMany(Star))
   .Mixin(Timestamps)
   .Raw(`@@map("comments")`);
 
-export default [Role, User, Post];
+// prettier-ignore
+Star
+  .Field('id',          Int(Id, Default('autoincrement()')))
+  .Field('postId',      Int(Nullable))
+  .Relation('post',     ManyToOne(Post, Fields('postId'), References('id')))
+  .Mixin(Timestamps)
+
+export default [Role, User, Post, Star];
+
+// model User {
+//   id           Int     @id @default(autoincrement())
+//   name         String?
+//   writtenPosts Post[]  @relation("WrittenPosts")
+//   pinnedPost   Post?   @relation("PinnedPost")
+// }
+//
+// model Post {
+//   id         Int     @id @default(autoincrement())
+//   title      String?
+//   author     User    @relation("WrittenPosts", fields: [authorId], references: [id])
+//   authorId   Int
+//   pinnedBy   User?   @relation(name: "PinnedPost", fields: [pinnedById], references: [id])
+//   pinnedById Int?
+//   @@map("comments")
+// }
+
+// let x = OneToOne(Post, 'WrittenPosts', Fields('wow'), References('wee'));
+// let a = OneToOne(Post, Fields('bestPostId'), References('id')); // good
+// let b = OneToOne(Post, References('id'), Fields('bestPostId')); // bad
+// let c = OneToOne(Post, Fields('bestPostId')); // bad
+// let d = OneToOne(Post); // good